+++ src/shared/game/core/index.ts
@@ -1,4 +1,0 @@
-export * from './map-renderer'
-export * from './iso-renderer'
-export * from './map-generator'
-export * from './camera'