+++ src/shared/game/index.ts
@@ -1,4 +1,0 @@
-export * from './core'
-export * from './constants'
-export * from './types'
-export * from './utils'