+++ src/shared/game/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './noise'
-export * from './texture-loader'
-export * from './tile-factory'